requirements.txt
absl-py==0.14.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.12.5
chamferdist==0.3.0
chardet==4.0.0
chumpy==0.70
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
freetype-py==2.2.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
h5py==3.2.1
idna==2.10
imageio==2.9.0
importlib-metadata==4.8.1
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
networkx==2.5.1
neural-renderer-pytorch==1.1.3
nibabel==3.2.1
numpy==1.20.1
oauthlib==3.1.1
opencv-python==4.5.1.48
packaging==21.0
Pillow==8.1.2
pointnet2-ops==3.0.0
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyglet==1.5.18
PyOpenGL==3.1.0
pyparsing==2.4.7
PyQt5==5.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrender==0.1.45
python-dateutil==2.8.1
PyWavelets==1.1.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.2.0
sip==4.19.8
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tifffile==2021.4.8
tornado==6.1
tqdm==4.60.0
transforms3d==0.3.1
trimesh==3.9.26
typing-extensions==3.10.0.2
urllib3==1.26.3
visdom==0.1.8.9
websocket-client==0.58.0
Werkzeug==2.0.2
zipp==3.6.0
